Several changes in this commit:
1. Add some fairness to scheduling.

When an lthread uses lthread_* functions more than 5 times
consecutively, then we'll yield it and let other lthreads to run
before we come back to it and resume it. This guarantees fairness
from lthread's perspective. This change also required that we queue
lthreads we'll resume instead of adding them to a link list. Since
we'll be yielding them to the back of the queue and resuming the
ones in the front.

2. Add lthread_pipe() which is equivalent to pipe(2) but it sets
the fds to non-blocking.

3. Add lthread_read and lthread_write to work with pipes. recv(2)
and send(2) cannot be used with pipes of course.

4. Add lthread_recv_exact() and lthread_read_exact(). These are
convenience functions to block an lthread until exact bytes are
read from socket/pipe.